# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sisCotacao — API",
    "description": "Contrato dos endpoints principais do ERP (fase 1). Autenticação por Bearer token emitido em /api/login (TTL 7 dias, revogável por logout ou troca de senha). Portal do fornecedor usa token próprio na URL e está fora deste contrato.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem (nginx proxy /api)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/plano-contas": {
      "get": {
        "tags": ["financeiro"],
        "summary": "Lista o plano de contas com filtros de natureza e rateabilidade",
        "parameters": [
          {"name": "tipo", "in": "query", "schema": {"type": "string", "enum": ["receita", "despesa"]}},
          {"name": "natureza", "in": "query", "schema": {"type": "string"}},
          {"name": "rateavel", "in": "query", "schema": {"type": "boolean"}},
          {"name": "somente_ativos", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {"200": {"description": "Contas do plano"}, "400": {"$ref": "#/components/responses/ErroValidacao"}}
      },
      "post": {
        "tags": ["financeiro"],
        "summary": "Cria conta de receita ou despesa com classificação",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PlanoContaIn"}}}},
        "responses": {"201": {"description": "Conta criada"}, "400": {"$ref": "#/components/responses/ErroValidacao"}}
      }
    },
    "/api/plano-contas/{conta_id}/uso": {
      "get": {
        "tags": ["financeiro"],
        "summary": "Consulta uso histórico de uma conta do plano",
        "parameters": [{"$ref": "#/components/parameters/ContaId"}],
        "responses": {"200": {"description": "Uso da conta"}, "404": {"description": "Conta não encontrada"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "HMAC (base64.payload.signature), TTL 7 dias"
      }
    },
    "parameters": {
...
      "ContaId": {
        "name": "conta_id",
        "in": "path",
        "required": true,
        "schema": {"type": "integer", "format": "int64"}
      },
...
    },
    "responses": {
...
      "ErroValidacao": {
        "description": "Filtros inválidos",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Erro"}
          }
        }
      }
    },
    "schemas": {
      "Erro": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "db_indisponivel"
          }
        }
      },
...
      "PlanoContaIn": {
        "type": "object",
        "required": ["codigo", "nome", "tipo"],
        "properties": {
          "codigo": {"type": "string"},
          "nome": {"type": "string"},
          "tipo": {"type": "string", "enum": ["receita", "despesa"]},
          "pai_id": {"type": "integer", "nullable": true},
          "natureza_custo": {"type": "string"},
          "politica_rateio": {"type": "string"},
          "exige_centro_custo": {"type": "boolean"},
          "exige_competencia": {"type": "boolean"},
          "permite_rateio": {"type": "boolean"},
          "componente_variavel": {"type": "string", "nullable": true}
        }
      },
...
    }
  }
}